+++ Core/SelfServicePaymentResult.cs
@@ -3,6 +3,6 @@ namespace Infrastructure.Services.Payment
 	public class SelfServicePaymentResult
 	{
 		public decimal AmountCharged { get; set; }
-		public double Discount { get; set; }
+		public decimal Discount { get; set; }
 	}
 }
